@@ -11,19 +11,20 @@
         }
     ],
     "require": {
-        "php": "^5.5.9 || ^7.0",
+        "php": "^5.6 || ^7.0",
         "illuminate/support": "5.1.* || 5.2.* || 5.3.*",
         "elasticsearch/elasticsearch": "^2.2",
         "mockery/mockery": "^0.9.5"
     },
     "require-dev": {
         "phpunit/phpunit": "^4.8 || ^5.0",
-        "orchestra/testbench": "^3.2",
-        "codeception/aspect-mock": "^1.0",
+        "orchestra/testbench": "^3.1",
+        "codeception/aspect-mock": "^1.0.0",
         "codeclimate/php-test-reporter": "dev-master"
     },
     "autoload": {
         "psr-4": {
+            "App\\": "app/",
             "Datashaman\\ElasticModel\\": "src"
         }
     },
