@@ -2,20 +2,29 @@
 
 use ArrayAccess;
 use Datashaman\ElasticModel\ArrayDelegate;
+use Illuminate\Support\Collection;
 
 class Results extends Base implements ArrayAccess
 {
     use ArrayDelegate;
 
-    protected $_arrayDelegate = 'results';
+    protected static $arrayDelegate = 'results';
 
     public function __get($name)
     {
         switch ($name) {
         case 'results':
-            return array_map(function ($hit) {
-                return new Result($hit);
-            }, $this->response->response['hits']['hits']);
+            $results = (new Collection($this->response->response['hits']['hits']))
+                ->map(function ($hit) {
+                    return new Result($hit);
+                });
+            return $results;
+        default:
+            return parent::__get($name);
         }
+    }
+
+    public function __call($name, $args) {
+        return call_user_func_array([ $this->results, $name ], $args);
     }
 }
